panning animation for the map
