+++ local_lens_web/src/components/SidePanel.jsx
@@ -1,15 +1,219 @@
 import { useState } from "react";
 
-function SidePanel() {
+function SidePanel({ map }) {
+  const [searchTerm, setSearchTerm] = useState("");
   const [activeTab, setActiveTab] = useState("places");
+  const [showSearchResults, setShowSearchResults] = useState(false);
+
+  const [nearbyPlaceSearch, setNearbySearch] = useState(null);
+  const [nearbyGebetaSearch, setNearbyGebeta] = useState(null);
+
+  const [navigateCoord, setNavigateCoor] = useState();
 
   const placeholderPlaces = [
-    { name: "Coffee Shop", distance: "0.2 km" },
-    { name: "Park", distance: "0.5 km" },
-    { name: "Grocery Store", distance: "0.8 km" },
-    { name: "Restaurant", distance: "1.2 km" },
-    { name: "Gym", distance: "1.5 km" },
+    { name: "Coffee Shop", distance: "0.2 km", reviewed: true, rating: 4.5 },
+    { name: "Park", distance: "0.5 km", reviewed: false },
+    { name: "Grocery Store", distance: "0.8 km", reviewed: true, rating: 3.8 },
+    { name: "Restaurant", distance: "1.2 km", reviewed: true, rating: 4.2 },
+    { name: "Gym", distance: "1.5 km", reviewed: false },
+    { name: "Bookstore", distance: "0.7 km", reviewed: true, rating: 4.0 },
+    { name: "Cinema", distance: "2.0 km", reviewed: false },
+    { name: "Bakery", distance: "0.3 km", reviewed: true, rating: 4.7 },
   ];
+
+  const filteredPlaces = placeholderPlaces.filter((place) =>
+    place.name.toLowerCase().includes(searchTerm.toLowerCase())
+  );
+
+  const reviewedPlaces = filteredPlaces.filter((place) => place.reviewed);
+  const unreviewedPlaces = filteredPlaces.filter((place) => !place.reviewed);
+
+  const fetchGebetaPlaces = async (term) => {
+    setNearbyGebeta(null);
+    console.log("This is your token", process.env.token);
+    try {
+      const response = await fetch(
+        `https://mapapi.gebeta.app/api/v1/route/geocoding?name=${term}&apiKey=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJjb21wYW55bmFtZSI6ImhhY2thdGhvbiIsImRlc2NyaXB0aW9uIjoiN2U0NmI2OTUtZmE5Yi00NDQ3LTliMjgtOGQ1NzQwMWQ3MjYwIiwiaWQiOiI4NzQzYTUyZC1jNGRjLTQyNjUtYWI1ZC0yMTJkNmI3NmVmNTgiLCJ1c2VybmFtZSI6ImhhY2F0aG9uIn0.qBPwmYTCHTjOQbJKFzFTCalyRdT_3fIW0BcqVno17g8`,
+        {
+          method: "GET",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+
+      if (!response.ok) {
+        throw new Error(`HTTP error! status: ${response.status}`);
+      }
+
+      const data = await response.json().then((data) => {
+        setNearbyGebeta(data.data);
+      });
+
+      console.log("Success:", data);
+    } catch (error) {
+      console.error("Error:", error);
+    }
+  };
+
+  const renderPlaces = (places) => {
+    return places.map((place, index) => (
+      <li key={index} className="mb-2 p-2 bg-gray-100 rounded">
+        <p className="font-semibold">{place.name}</p>
+        <p className="text-sm text-gray-600">Distance: {place.distance}</p>
+        {place.reviewed && (
+          <p className="text-sm text-blue-600">Rating: {place.rating}</p>
+        )}
+      </li>
+    ));
+  };
+  const renderUnreviewed = (places) => {
+    return places == null ? (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 100 100"
+        preserveAspectRatio="xMidYMid"
+        style={{
+          shapeRendering: "auto",
+          display: "block",
+          background: "transparent",
+        }}
+        width="100%"
+        height="200"
+      >
+        <g>
+          <circle fill="#464646" r="10" cy="50" cx="84">
+            <animate
+              begin="0s"
+              keySplines="0 0.5 0.5 1"
+              values="10;0"
+              keyTimes="0;1"
+              calcMode="spline"
+              dur="0.25s"
+              repeatCount="indefinite"
+              attributeName="r"
+            ></animate>
+            <animate
+              begin="0s"
+              values="#464646;#464646;#dfdfdf;#a0a0a0;#464646"
+              keyTimes="0;0.25;0.5;0.75;1"
+              calcMode="discrete"
+              dur="1s"
+              repeatCount="indefinite"
+              attributeName="fill"
+            ></animate>
+          </circle>
+          <circle fill="#464646" r="10" cy="50" cx="16">
+            <animate
+              begin="0s"
+              keySplines="0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1"
+              values="0;0;10;10;10"
+              keyTimes="0;0.25;0.5;0.75;1"
+              calcMode="spline"
+              dur="1s"
+              repeatCount="indefinite"
+              attributeName="r"
+            ></animate>
+            <animate
+              begin="0s"
+              keySplines="0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1"
+              values="16;16;16;50;84"
+              keyTimes="0;0.25;0.5;0.75;1"
+              calcMode="spline"
+              dur="1s"
+              repeatCount="indefinite"
+              attributeName="cx"
+            ></animate>
+          </circle>
+          <circle fill="#a0a0a0" r="10" cy="50" cx="50">
+            <animate
+              begin="-0.25s"
+              keySplines="0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1"
+              values="0;0;10;10;10"
+              keyTimes="0;0.25;0.5;0.75;1"
+              calcMode="spline"
+              dur="1s"
+              repeatCount="indefinite"
+              attributeName="r"
+            ></animate>
+            <animate
+              begin="-0.25s"
+              keySplines="0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1"
+              values="16;16;16;50;84"
+              keyTimes="0;0.25;0.5;0.75;1"
+              calcMode="spline"
+              dur="1s"
+              repeatCount="indefinite"
+              attributeName="cx"
+            ></animate>
+          </circle>
+          <circle fill="#dfdfdf" r="10" cy="50" cx="84">
+            <animate
+              begin="-0.5s"
+              keySplines="0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1"
+              values="0;0;10;10;10"
+              keyTimes="0;0.25;0.5;0.75;1"
+              calcMode="spline"
+              dur="1s"
+              repeatCount="indefinite"
+              attributeName="r"
+            ></animate>
+            <animate
+              begin="-0.5s"
+              keySplines="0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1"
+              values="16;16;16;50;84"
+              keyTimes="0;0.25;0.5;0.75;1"
+              calcMode="spline"
+              dur="1s"
+              repeatCount="indefinite"
+              attributeName="cx"
+            ></animate>
+          </circle>
+          <circle fill="#464646" r="10" cy="50" cx="16">
+            <animate
+              begin="-0.75s"
+              keySplines="0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1"
+              values="0;0;10;10;10"
+              keyTimes="0;0.25;0.5;0.75;1"
+              calcMode="spline"
+              dur="1s"
+              repeatCount="indefinite"
+              attributeName="r"
+            ></animate>
+            <animate
+              begin="-0.75s"
+              keySplines="0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1;0 0.5 0.5 1"
+              values="16;16;16;50;84"
+              keyTimes="0;0.25;0.5;0.75;1"
+              calcMode="spline"
+              dur="1s"
+              repeatCount="indefinite"
+              attributeName="cx"
+            ></animate>
+          </circle>
+          <g></g>
+        </g>
+      </svg>
+    ) : (
+      places.map((place, index) => (
+        <li
+          key={index}
+          className="mb-2 p-2 bg-gray-100 rounded cursor-pointer hover:bg-gray-300 transition-colors"
+          onClick={ () => map.setView([place.latitude, place.longitude], 20, { animate: true, duration: 2})}
+        >
+          <p className="font-semibold">{place.name}</p>
+          <p className="text-sm text-gray-600">Distance: 1.0km</p>
+          <p className="text-sm text-blue-600">Type: {place.type}</p>
+        </li>
+      ))
+    );
+  };
+
+  const handleSearchChange = (e) => {
+    fetchGebetaPlaces(e.target.value);
+    setSearchTerm(e.target.value);
+    setShowSearchResults(e.target.value.length > 0);
+  };
 
   const placeholderPeople = [
     { name: "John Doe", distance: "0.3 km" },
@@ -56,7 +260,27 @@ function SidePanel() {
 
   return (
     <div className="rounded-lg w-1/4 bg-white shadow-2xl absolute right-7 top-28 z-10">
-      <div className="rounded-lg bg-white p-4 overflow-y-auto h-full flex flex-col">
+      <div
+        className="rounded-lg bg-white p-4 overflow-y-auto h-full flex flex-col"
+        style={{ maxHeight: "80dvh" }}
+      >
+        <input
+          type="text"
+          placeholder="Search places..."
+          value={searchTerm}
+          onChange={handleSearchChange}
+          className="w-full p-2 mb-4 border rounded"
+        />
+        {showSearchResults && (
+          <div className="absolute top-16 left-0 right-0 bg-white border rounded shadow-lg z-10 max-h-96 overflow-y-auto">
+            <div className="p-4">
+              <h3 className="text-lg font-bold mb-2">Reviewed Places</h3>
+              <ul className="mb-4">{renderPlaces(reviewedPlaces)}</ul>
+              <h3 className="text-lg font-bold mb-2">Unreviewed Places</h3>
+              <ul>{renderUnreviewed(nearbyGebetaSearch)}</ul>
+            </div>
+          </div>
+        )}
         <div className="flex mb-4">
           <button
             className={`flex-1 py-2 ${
